requirements.txt
mysql-connector-python
pandas
streamlit